+++ Blobz/Machines.java
@@ -19,7 +19,7 @@ public abstract class Machines extends Actor
     protected GreenfootSound[] place = new GreenfootSound[20];
     public GreenfootSound placeSound;
     protected int effectsIndex = 0, placeIndex = 0;
-    protected int value;
+    protected double value;
     
     /**
      * Allows an image of the seleted machine to follow mouse cursor
